refactor: 예약 가능한 시간 목록 조회 기능 reservation으로 이동
