+++ src/main/java/roomescape/console/repository/ReservationTimeConsoleRepository.java
@@ -6,7 +6,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
 import roomescape.core.domain.ReservationTime;
-import roomescape.core.dto.BookingTimeResponseDto;
 import roomescape.core.repository.ReservationTimeRepository;
 
 public class ReservationTimeConsoleRepository implements ReservationTimeRepository {
@@ -31,11 +30,6 @@ public class ReservationTimeConsoleRepository implements ReservationTimeReposito
     @Override
     public List<ReservationTime> findAll() {
         return Collections.unmodifiableList(reservationTimes);
-    }
-
-    @Override
-    public List<BookingTimeResponseDto> findAllByDateAndThemeId(final String date, final long themeId) {
-        return List.of();
     }
 
     @Override
