+++ src/main/java/roomescape/core/service/ReservationTimeService.java
@@ -4,7 +4,6 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.core.domain.ReservationTime;
-import roomescape.core.dto.BookingTimeResponseDto;
 import roomescape.core.dto.ReservationTimeRequestDto;
 import roomescape.core.dto.ReservationTimeResponseDto;
 import roomescape.core.repository.ReservationRepository;
@@ -43,11 +42,6 @@ public class ReservationTimeService {
                 .stream()
                 .map(ReservationTimeResponseDto::new)
                 .toList();
-    }
-
-    @Transactional(readOnly = true)
-    public List<BookingTimeResponseDto> findBookable(final String date, final long themeId) {
-        return reservationTimeRepository.findAllByDateAndThemeId(date, themeId);
     }
 
     @Transactional
