+++ src/main/java/roomescape/web/repository/ReservationTimeRepositoryImpl.java
@@ -10,7 +10,6 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.core.domain.ReservationTime;
-import roomescape.core.dto.BookingTimeResponseDto;
 import roomescape.core.repository.ReservationTimeRepository;
 
 @Repository
@@ -35,29 +34,6 @@ public class ReservationTimeRepositoryImpl implements ReservationTimeRepository 
     @Override
     public List<ReservationTime> findAll() {
         return jdbcTemplate.query("SELECT id, start_at FROM reservation_time", getReservationTimeRowMapper());
-    }
-
-    @Override
-    public List<BookingTimeResponseDto> findAllByDateAndThemeId(final String date, final long themeId) {
-        final String query = """
-            SELECT t.id, t.start_at, r.id IS NOT NULL AS already_booked
-            FROM reservation_time AS t 
-            LEFT JOIN (
-                SELECT * FROM reservation WHERE date LIKE ? AND theme_id = ?
-            ) AS r
-            ON t.id = r.time_id;
-            """;
-        return jdbcTemplate.query(query, getBookingTimeRowMapper(), date, themeId);
-    }
-
-    private RowMapper<BookingTimeResponseDto> getBookingTimeRowMapper() {
-        return (resultSet, rowNum) -> {
-            final Long timeId = resultSet.getLong("id");
-            final String timeValue = resultSet.getString("start_at");
-            final boolean alreadyBooked = resultSet.getBoolean("already_booked");
-
-            return new BookingTimeResponseDto(timeId, timeValue, alreadyBooked);
-        };
     }
 
     @Override
